+++ src/utils.ts
@@ -1,6 +1,10 @@
+/* eslint-disable no-loop-func */
+/* eslint-disable no-await-in-loop */
 import http from 'http'
-import { findIndex, isNil, mapValues, omitBy } from 'lodash'
+import { findIndex, mapValues } from 'lodash'
 import { CoreV1Api, KubeConfig, V1Secret, V1ObjectMeta, V1ServiceAccount } from '@kubernetes/client-node'
+import retry, { Options } from 'async-retry'
+import fetch, { RequestInit } from 'node-fetch'
 
 let apiClient: CoreV1Api
 
@@ -95,7 +99,7 @@ export async function doApiCall(
   }
 }
 
-export function handleErrors(errors: string[]) {
+export function handleErrors(errors: string[]): void {
   if (errors.length) {
     console.error(`Errors found: ${JSON.stringify(errors, null, 2)}`)
     process.exit(1)
@@ -186,10 +190,44 @@ export async function deletePullSecret(teamId: string, name: string): Promise<vo
   }
 }
 
-export function cleanValues(inObj) {
-  const obj = omitBy(inObj, isNil)
-  Object.keys(obj).forEach((k) => {
-    if (typeof obj[k] === 'object') obj[k] = cleanValues(obj[k])
-  })
-  return obj
+export async function waitTillAvailable(url: string, status = 200): Promise<void> {
+  const retryOptions: Options = {
+    retries: 10,
+    factor: 2,
+    // minTimeout: The number of milliseconds before starting the first retry. Default is 1000.
+    minTimeout: 1000,
+    // The maximum number of milliseconds between two retries.
+    maxTimeout: 30000,
+  }
+  const minimumSuccessful = 10
+  let count = 0
+  const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms))
+  try {
+    do {
+      console.log('retry count: ', count)
+      await retry(async (bail) => {
+        try {
+          const fetchOptions: RequestInit = {
+            redirect: 'follow',
+          }
+          const res = await fetch(url, fetchOptions)
+          if (res.status !== status) {
+            console.warn(`GET ${res.url} ${res.status}`)
+            bail(new Error(`Retry`))
+          } else {
+            count += 1
+            await delay(1000)
+          }
+        } catch (e) {
+          // Print system errors like ECONNREFUSED
+          console.error(e.message)
+          count = 0
+          throw e
+        }
+      }, retryOptions)
+    } while (count < minimumSuccessful)
+  } catch (e) {
+    console.error('Max retry tries has been reached: ', e)
+    process.exit(1)
+  }
 }